refactor(runbotics-desktop): :recycle: fix after CR
