+++ runbotics/runbotics-desktop/src/action/file/file.action-handler.ts
@@ -1,52 +1,9 @@
-import { StatelessActionHandler, DesktopRunRequest } from '@runbotics/runbotics-sdk';
+import { StatelessActionHandler } from '@runbotics/runbotics-sdk';
 import fs from 'fs';
-import path from 'path';
 import { Injectable } from '@nestjs/common';
-import { FileAction } from 'runbotics-common';
-
-export type FileActionRequest =
-    | DesktopRunRequest<FileAction.WRITE_FILE, FileWriteFileActionInput>
-    | DesktopRunRequest<FileAction.READ_FILE, FileReadFileActionInput>
-    | DesktopRunRequest<FileAction.APPEND_FILE, FileAppendFileActionInput>
-    | DesktopRunRequest<FileAction.CREATE_FILE, FileCreateFileActionInput>
-    | DesktopRunRequest<FileAction.REMOVE_FILE, FileRemoveFileActionInput>;
-
-// ----
-export type FileAppendFileActionInput = {
-    content: string;
-    path: string;
-    separator: string;
-};
-export type FileAppendFileActionOutput = any;
-
-// ----
-export type FileCreateFileActionInput = {
-    path: string;
-    conflict: 'Overwrite' | 'Extend name' | 'Throw Error';
-    // fileName: string;
-};
-export type FileCreateFileActionOutput = any;
-
-// ----
-export type FileRemoveFileActionInput = {
-    path: string;
-};
-export type FileRemoveFileActionOutput = any;
-
-// ----
-export type FileReadFileActionInput = {
-    path: string;
-};
-export type FileReadFileActionOutput = any;
-
-// ----
-export type FileWriteFileActionInput = {
-    path: string;
-    content: string;
-};
-export type FileWriteFileActionOutput = any;
-
-
+import { FileActionRequest, FileAppendFileActionInput, FileAppendFileActionOutput, FileCreateFileActionInput, FileCreateFileActionOutput, FileReadFileActionInput, FileReadFileActionOutput, FileRemoveFileActionInput, FileRemoveFileActionOutput, FileWriteFileActionInput, FileWriteFileActionOutput } from './types';
+import { createNewFile, getUniqueFileName } from './file.utils';
+import { ConflictFile } from 'runbotics-common';
 @Injectable()
 export default class FileActionHandler extends StatelessActionHandler {
     constructor() {
@@ -65,15 +22,15 @@ export default class FileActionHandler extends StatelessActionHandler {
     async createFile(input: FileCreateFileActionInput): Promise<FileCreateFileActionOutput> {
         if(fs.existsSync(input.path)) {
             switch (input.conflict) {
-                case 'Overwrite': 
+                case ConflictFile.OVERWRITE: 
                     await createNewFile(input.path);
                     return 'File overwritten successfully';
-                case 'Extend name': {
+                case ConflictFile.EXTEND_NAME: {
                     const newFilePath = getUniqueFileName(input.path);
                     await createNewFile(newFilePath);
                     return 'File with extended name created successfully';
                 }
-                case 'Throw Error':
+                case ConflictFile.THROW_ERROR:
                     throw Error('File with the same name already exists');
                 default: 
                     await createNewFile(input.path);
@@ -130,27 +87,3 @@ export default class FileActionHandler extends StatelessActionHandler {
     }
 }
 
-async function createNewFile(path: string){
-    try {
-        fs.closeSync(fs.openSync(`${path}`, 'a'));
-    } catch (err) {
-        throw Error(err);
-    }
-}
-
-function getUniqueFileName(filePath) {
-    const dir = path.dirname(filePath);
-    const ext = path.extname(filePath);
-    const baseName = path.basename(filePath, ext).replace(/\s\(\d+\)/, '');
-    
-    let counter = 1;
-    let newFilePath = path.join(dir, `${baseName} (${counter})${ext}`);
-    
-    // Keep increasing counter until we find a free name
-    while (fs.existsSync(newFilePath)) {
-        counter++;
-        newFilePath = path.join(dir, `${baseName} (${counter})${ext}`);
-    }
-
-    return newFilePath;
-}
