+++ runbotics/runbotics-desktop/src/action/file/index.ts
@@ -1,2 +1,3 @@
 export * from './file.action-handler';
+export * from './types';
 export { default } from './file.action-handler';
